upate 08.08.2019

+++ src/main/java/com/levermann/entityclass/Levermannschritte.java
@@ -25,49 +25,49 @@ import java.io.Serializable;
 
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private Integer id;
+    private Float id;
 
     @Column(name = "Unternehmenname_Levermannschritte", unique = true, nullable = true)
     private String Unternehmenname_Levermannschritte;
 
     @Column(name = "Eigenkapitalrendite", unique = true, nullable = true)
-    private Integer Eigenkapitalrendite;
+    private Float Eigenkapitalrendite;
 
     @Column(name = "EBITMarge", unique = true, nullable = true)
-    private Integer EBITMarge;
+    private Float EBITMarge;
 
     @Column(name = "Eigenkapitalquote", unique = true, nullable = true)
-    private Integer Eigenkapitalquote;
+    private Float Eigenkapitalquote;
 
     @Column(name = "KursGewinnVerhaeltnis", unique = true, nullable = true)
-    private Integer KursGewinnVerhaeltnis;
+    private Float KursGewinnVerhaeltnis;
 
     @Column(name = "KursGewinnVerhaeltnisAktuell", unique = true, nullable = true)
-    private Integer KursGewinnVerhaeltnisAktuell;
+    private Float KursGewinnVerhaeltnisAktuell;
 
     @Column(name = "Analystenmeinungen", unique = true, nullable = true)
-    private Integer Analystenmeinungen;
+    private Float Analystenmeinungen;
 
     @Column(name = "ReaktionaufQuartalszahlen", unique = true, nullable = true)
-    private Integer ReaktionaufQuartalszahlen;
+    private Float ReaktionaufQuartalszahlen;
 
     @Column(name = "Gewinnrevision", unique = true, nullable = true)
-    private Integer Gewinnrevision;
+    private Float Gewinnrevision;
 
     @Column(name = "Kursverlauf6Monate", unique = true, nullable = true)
-    private Integer Kursverlauf6Monate;
+    private Float Kursverlauf6Monate;
 
     @Column(name = "Kursverlauf12Monate", unique = true, nullable = true)
-    private Integer Kursverlauf12Monate;
+    private Float Kursverlauf12Monate;
 
     @Column(name = "Kursmomentum", unique = true, nullable = true)
-    private Integer Kursmomentum;
+    private Float Kursmomentum;
 
     @Column(name = "Dreimonatsreversal", unique = true, nullable = true)
-    private Integer Dreimonatsreversal;
+    private Float Dreimonatsreversal;
 
     @Column(name = "Gewinnwachstum", unique = true, nullable = true)
-    private Integer Gewinnwachstum;
+    private Float Gewinnwachstum;
 
 
     public String getLevermannschrittAnalyseNameId() {
@@ -86,118 +86,118 @@ import java.io.Serializable;
         Unternehmenname_Levermannschritte = unternehmenname_Levermannschritte;
     }
 
-    public Integer getEigenkapitalrendite() {
+    public Float getEigenkapitalrendite() {
         return Eigenkapitalrendite;
     }
 
-    public void setEigenkapitalrendite(Integer eigenkapitalrendite) {
+    public void setEigenkapitalrendite(Float eigenkapitalrendite) {
         Eigenkapitalrendite = eigenkapitalrendite;
     }
 
-    public Integer getEBITMarge() {
+    public Float getEBITMarge() {
         return EBITMarge;
     }
 
-    public void setEBITMarge(Integer EBITMarge) {
+    public void setEBITMarge(Float EBITMarge) {
         this.EBITMarge = EBITMarge;
     }
 
-    public Integer getEigenkapitalquote() {
+    public Float getEigenkapitalquote() {
         return Eigenkapitalquote;
     }
 
-    public void setEigenkapitalquote(Integer eigenkapitalquote) {
+    public void setEigenkapitalquote(Float eigenkapitalquote) {
         Eigenkapitalquote = eigenkapitalquote;
     }
 
-    public Integer getKursGewinnVerhaeltnis() {
+    public Float getKursGewinnVerhaeltnis() {
         return KursGewinnVerhaeltnis;
     }
 
-    public void setKursGewinnVerhaeltnis(Integer kursGewinnVerhaeltnis) {
+    public void setKursGewinnVerhaeltnis(Float kursGewinnVerhaeltnis) {
         KursGewinnVerhaeltnis = kursGewinnVerhaeltnis;
     }
 
-    public Integer getKursGewinnVerhaeltnisAktuell() {
+    public Float getKursGewinnVerhaeltnisAktuell() {
         return KursGewinnVerhaeltnisAktuell;
     }
 
-    public void setKursGewinnVerhaeltnisAktuell(Integer kursGewinnVerhaeltnisAktuell) {
+    public void setKursGewinnVerhaeltnisAktuell(Float kursGewinnVerhaeltnisAktuell) {
         KursGewinnVerhaeltnisAktuell = kursGewinnVerhaeltnisAktuell;
     }
 
-    public Integer getAnalystenmeinungen() {
+    public Float getAnalystenmeinungen() {
         return Analystenmeinungen;
     }
 
-    public void setAnalystenmeinungen(Integer analystenmeinungen) {
+    public void setAnalystenmeinungen(Float analystenmeinungen) {
         Analystenmeinungen = analystenmeinungen;
     }
 
-    public Integer getReaktionaufQuartalszahlen() {
+    public Float getReaktionaufQuartalszahlen() {
         return ReaktionaufQuartalszahlen;
     }
 
-    public void setReaktionaufQuartalszahlen(Integer reaktionaufQuartalszahlen) {
+    public void setReaktionaufQuartalszahlen(Float reaktionaufQuartalszahlen) {
         ReaktionaufQuartalszahlen = reaktionaufQuartalszahlen;
     }
 
-    public Integer getGewinnrevision() {
+    public Float getGewinnrevision() {
         return Gewinnrevision;
     }
 
-    public void setGewinnrevision(Integer gewinnrevision) {
+    public void setGewinnrevision(Float gewinnrevision) {
         Gewinnrevision = gewinnrevision;
     }
 
-    public Integer getKursverlauf6Monate() {
+    public Float getKursverlauf6Monate() {
         return Kursverlauf6Monate;
     }
 
-    public void setKursverlauf6Monate(Integer kursverlauf6Monate) {
+    public void setKursverlauf6Monate(Float kursverlauf6Monate) {
         Kursverlauf6Monate = kursverlauf6Monate;
     }
 
-    public Integer getKursverlauf12Monate() {
+    public Float getKursverlauf12Monate() {
         return Kursverlauf12Monate;
     }
 
-    public void setKursverlauf12Monate(Integer kursverlauf12Monate) {
+    public void setKursverlauf12Monate(Float kursverlauf12Monate) {
         Kursverlauf12Monate = kursverlauf12Monate;
     }
 
-    public Integer getKursmomentum() {
+    public Float getKursmomentum() {
         return Kursmomentum;
     }
 
-    public void setKursmomentum(Integer kursmomentum) {
+    public void setKursmomentum(Float kursmomentum) {
         Kursmomentum = kursmomentum;
     }
 
-    public Integer getDreimonatsreversal() {
+    public Float getDreimonatsreversal() {
         return Dreimonatsreversal;
     }
 
-    public void setDreimonatsreversal(Integer dreimonatsreversal) {
+    public void setDreimonatsreversal(Float dreimonatsreversal) {
         Dreimonatsreversal = dreimonatsreversal;
     }
 
-    public Integer getGewinnwachstum() {
+    public Float getGewinnwachstum() {
         return Gewinnwachstum;
     }
 
-    public void setGewinnwachstum(Integer gewinnwachstum) {
+    public void setGewinnwachstum(Float gewinnwachstum) {
         Gewinnwachstum = gewinnwachstum;
     }
 
     public Levermannschritte() {
     }
 
-    public Integer getId() {
+    public Float getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Float id) {
         this.id = id;
     }
 }
